+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for Laravel 9.x and below
 - Support for Lumen
+- Custom validation error placeholders